drm/gud: Add functionality for the USB gadget side
Since the USB gadget/device has to reach into the DRM internals, part of
the code is placed in the DRM subsystem as a separate module. All calls
into this module runs in process context and are serialized, except one
function that runs in interrupt context.

Since both the gadget side and the DRM side can disable the gadget,
special care has been taken to make that safe.

Changelog:

20210522:
  Fix 64-bit build warning

20210601:
  Use async flushing, speeds up SPI panels.
  Add configfs filters for: compression, formats, connectors

Signed-off-by: Noralf Trønnes <noralf@tronnes.org>