arm64: pinenote_defconfig: Improve SCMI configuration
The SCMI specification provides a standardized interface for power,
performance and resource management on a SoC.
The low-level management actions are performed by a system controller
that directly controls the SoC hardware or platform.
The system controller provides SCMI interfaces to its clients.
A typical example of an SCMI client is an Operating System kernel.

To enable basic support for that:
- Explicitly make ``ARM_SCMI_TRANSPORT_MAILBOX``, ``ARM_SCMI_POWER_DOMAIN``
  and ``ARM_SCMI_PERF_DOMAIN ``  builtin to make sure they're enabled even
  if upstream changes the default away from 'y'
- Make the SCMI settings related to power and pinctrl builtin as they
  can be needed for (initial) power operations and early boot
- Build the other modules for the basic SCMI framework as modules

This makes the SCMI configuration in line with what was used when
testing the TF-A "feat(rk3568): support SCMI for clock/reset domain"
patch set.

Link: https://developer.arm.com/documentation/102886/001/
Link: https://developer.arm.com/documentation/den0056/latest/
Link: https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/31265