arm64: pinenote_defconfig: Update devfreq configuration
The use of 'Update' is a bit of an understatement as the configuration
is completely changed ... because the old one did not make any sense.

The problem:
The performance and powersave governors only set it to max and min
respectively, which is rarely useful or a good idea, so remove those.
The ``ARM_RK3399_DMC_DEVFREQ`` module is really only possibly useful on
rk3399 based devices, but apparently it does not work there either, so
remove that option as well.

The solution:
Replace the configuration with what's used on Debian, or at least the
parts that are or could be useful for the PineNote.

First explicitly enable the generic DVFS support as that is useful and
important (and effectively enabled in the current PN kernel).

Secondly, make the SIMPLE_ONDEMAND governor built-in as generally a/the
ondemand governor is useful. Furthermore it was made built-in on Debian
as well as the ARM Mali GPU driver needs a devfreq governor and the
SIMPLE_ONDEMAND is excellent for that.

Also make the PASSIVE governor available (as module) as that's also
available in the Debian kernel and it uses OPP tables (which are defined
in the DeviceTree).

Keep the USERSPACE governor, but make it a module. Not sure if it's
useful, but to reduce the risk of breaking things, don't remove it.

Make devfreq-event support explicitly available and enable the
``DEVFREQ_EVENT_ROCKCHIP_DFI`` driver as a module as that is
used/configured in the rk356x SoCs.