rk3566-pinenote.dtsi: set CPLL_333M to 33.33 MHZ and use in rockchip_ebc direct mode
This allows the panel to be run at the recommended 33.33 MHz sclk,
decreasing frame processing times by a bit more than a ms. This only
works in direct mode by setting EBC_DSP_CTRL_DSP_SDCLK_DIV to 0. This
behaviour is not documented in the Rockchip RK3568 TRM, but is used by
the Lenovo Smart Paper.

LUT mode and 3WIN mode are not compatible with this setting and need
to be run at sclk 25 MHz (dclk 200 MHz, clock divider of
8). Previously 31.25 MHz were supported by setting dclk to 250 MHz,
which resulted in CPLL_333M to be set to 250 MHz.