rockchip-ebc: introduce the globre_convert_before module parameter
This parameter, when set to 1, will lead to the final buffers being
converted to either A2 or DU4-compatible grayscale values when a global
refresh is triggered.
When the default_waveform == 1 (A2), the final buffer contents are
dithered.
When the default_waveform == 3 (DU4), then the final buffer contents are
quantizised into the four grayscale values that DU4 can work with.

Note that this conversion behavior is automatically triggered on the
first global refresh after coming out of suspend, as usually we want to
draw the offscreen content in GC16, but also want to make sure the
buffer content is compatible with the currently selected waveform.